requirements.txt
pygame==2.6.1
langchain==0.3.19
langchain-core==0.3.36
langchainhub==0.1.21
langchain-openai==0.3.0
Flask==3.1.0
streamlit==1.42.2
python-dotenv
